Revert "Checkpoint:  start work on generic "start-to-end" connection in"
This reverts commit 351dad139730b1b43721cfedfd54594044f1e518.

Why?  Because a clever Connection() method that takes any start/end
input may be able to produce an Arc/Segment or Arc/Arc pair that
satisfies the xy-plane geometry --- but then what is it going to
do for elevation and superelevation?

Perhaps user could specify a ZPoint to set a target for the joining
point wherever that is?  Not clear; revisit later.